[ins] wrapper for ins float_invariant
