+++ conf/firmwares/subsystems/fixedwing/ins_float_invariant.makefile
@@ -2,7 +2,7 @@
 
 # attitude and speed estimation for fixedwings via invariant filter
 
-INS_CFLAGS += -DAHRS_TYPE_H=\"subsystems/ins/ins_float_invariant.h\"
+INS_CFLAGS += -DAHRS_TYPE_H=\"subsystems/ins/ins_float_invariant_wrapper.h\"
 INS_CFLAGS += -DUSE_AHRS_ALIGNER
 INS_CFLAGS += -DUSE_AHRS
 INS_CFLAGS += -DINS_UPDATE_FW_ESTIMATOR
@@ -11,6 +11,7 @@ INS_SRCS += $(SRC_SUBSYSTEMS)/ahrs.c
 INS_SRCS += $(SRC_SUBSYSTEMS)/ahrs/ahrs_aligner.c
 INS_SRCS += $(SRC_SUBSYSTEMS)/ins.c
 INS_SRCS += $(SRC_SUBSYSTEMS)/ins/ins_float_invariant.c
+INS_SRCS += $(SRC_SUBSYSTEMS)/ins/ins_float_invariant_wrapper.c
 
 
 ifneq ($(AHRS_ALIGNER_LED),none)
