+++ conf/firmwares/subsystems/rotorcraft/ins_float_invariant.makefile
@@ -4,7 +4,7 @@
 
 USE_MAGNETOMETER ?= 1
 
-INS_CFLAGS += -DAHRS_TYPE_H=\"subsystems/ins/ins_float_invariant.h\"
+INS_CFLAGS += -DAHRS_TYPE_H=\"subsystems/ins/ins_float_invariant_wrapper.h\"
 INS_CFLAGS += -DUSE_AHRS_ALIGNER
 INS_CFLAGS += -DUSE_AHRS
 # for geo mag
@@ -18,7 +18,7 @@ INS_SRCS += $(SRC_SUBSYSTEMS)/ahrs.c
 INS_SRCS += $(SRC_SUBSYSTEMS)/ahrs/ahrs_aligner.c
 INS_SRCS += $(SRC_SUBSYSTEMS)/ins.c
 INS_SRCS += $(SRC_SUBSYSTEMS)/ins/ins_float_invariant.c
-
+INS_SRCS += $(SRC_SUBSYSTEMS)/ins/ins_float_invariant_wrapper.c
 
 ifneq ($(AHRS_ALIGNER_LED),none)
   INS_CFLAGS += -DAHRS_ALIGNER_LED=$(AHRS_ALIGNER_LED)
